+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+-  Support [Ipv6] format requested_host in --connect-to #197
+
 # 0.5.5 (2022-09-19)
 
 - Add colors to the tui view #64